Add logic to presentation layer

+++ data/src/main/java/com/pavelsemak/weatherapp/data/model/WeatherResponse.java
@@ -2,26 +2,26 @@ package com.pavelsemak.weatherapp.data.model;
 
 public class WeatherResponse {
 
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
     private WeatherItemResponse currently;
     private WeatherBlockResponse hourly;
     private WeatherBlockResponse daily;
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
